AI-ML-Internship-Eminds/RAG_App_using_langchain/requirements.txt
flask
python-dotenv
langchain
langchain-community
langchain-groq
chromadb